+++ tests/Tests/Parametizer/Autocompletion/scripts/deep-default-subcommand.php
@@ -8,14 +8,16 @@ require_once __DIR__ . '/../../../init-console.php';
 
 TestUtils::newConfig()
     ->newSubcommandSwitch('subcommand')
-    ->newSubcommand('test', TestUtils::newConfig())
     ->newSubcommand(
         'level-2',
         TestUtils::newConfig()
-            ->newSubcommandSwitch('subcommand')
-            ->newSubcommand('test-2', TestUtils::newConfig()),
+            ->newSubcommandSwitch('subcommand-l2')
+            ->newSubcommand(
+                'level-3',
+                TestUtils::newConfig()
+                    ->newArgument('argument')
+                    ->allowedValues(['asd', 'zxc', 'qwe']),
+            ),
     )
 
     ->run();
-
-throw new Exception('No built-in subcommand call');